refactor: 운동 생성 검증 로직 이동
